+++ src/HelpDiv/HelpDiv.spec.tsx
@@ -2,19 +2,19 @@ import React from 'react'
 import renderer from 'react-test-renderer'
 import 'jest-styled-components'
 
-import Tooltip from './index.js'
+import HelpDiv from './index.js'
 
-describe('Tooltip', () => {
+describe('HelpDiv', () => {
   test('default', () => {
-    const tree = renderer.create(<Tooltip>Some content</Tooltip>).toJSON()
+    const tree = renderer.create(<HelpDiv>Some content</HelpDiv>).toJSON()
     expect(tree).toMatchSnapshot()
   })
   test('overrides', () => {
     const tree = renderer
       .create(
-        <Tooltip width="28ch" background="#FFF" leftOrRight="left">
+        <HelpDiv width="28ch" background="#FFF" leftOrRight="left">
           Some content
-        </Tooltip>
+        </HelpDiv>
       )
       .toJSON()
     expect(tree).toMatchSnapshot()
