+++ src/index.ts
@@ -10,4 +10,4 @@ export * from './Heading/index.js'
 export { default as Section } from './Section/index.js'
 export { default as Dialog } from './Dialog/index.js'
 export { default as Link } from './Link/index.js'
-export { default as Tooltip } from './Tooltip/index.js'
+export { default as HelpDiv } from './HelpDiv/index.js'
